Completed the user and admin segerigation

+++ Backend/controllers/problemController.js
@@ -9,6 +9,7 @@ export const createProblem = async (req, res) => {
             title,
             statement,
             problemId: uuidv4(),
+            createdBy: req.user._id,
         });
 
         await newProblem.save();
@@ -19,25 +20,64 @@ export const createProblem = async (req, res) => {
 };
 
 export const getAllProblems = async (req, res) => {
-    const problems = await Problem.find().sort({ createdAt: 1 });
+    const problems = await Problem.find()
+        .populate("createdBy", "_id")
+        .sort({ createdAt: 1 });
     res.json(problems);
 };
 
 export const getProblemById = async (req, res) => {
-    const problem = await Problem.findOne({ problemId: req.params.id });
+    const problem = await Problem.findOne({
+        problemId: req.params.id,
+    }).populate("createdBy", "_id");
+
+    if (!problem) {
+        return res.status(404).json({ message: "Problem not found" });
+    }
+
     res.json(problem);
 };
 
 export const updateProblem = async (req, res) => {
-    const updated = await Problem.findOneAndUpdate(
-        { problemId: req.params.id },
-        req.body,
-        { new: true },
-    );
-    res.json(updated);
+    try {
+        const problem = await Problem.findOne({ problemId: req.params.id });
+
+        if (!problem) {
+            return res.status(404).json({ message: "Problem not found" });
+        }
+
+        if (problem.createdBy.toString() !== req.user._id.toString()) {
+            return res.status(403).json({ message: "Not authorized" });
+        }
+
+        const updated = await Problem.findOneAndUpdate(
+            { problemId: req.params.id },
+            req.body,
+            { new: true },
+        );
+
+        res.json(updated);
+    } catch (err) {
+        res.status(500).json({ error: err.message });
+    }
 };
 
 export const deleteProblem = async (req, res) => {
-    await Problem.findOneAndDelete({ problemId: req.params.id });
-    res.json({ message: "Deleted" });
+    try {
+        const problem = await Problem.findOne({ problemId: req.params.id });
+
+        if (!problem) {
+            return res.status(404).json({ message: "Problem not found" });
+        }
+
+        if (problem.createdBy.toString() !== req.user._id.toString()) {
+            return res.status(403).json({ message: "Not authorized" });
+        }
+
+        await Problem.findOneAndDelete({ problemId: req.params.id });
+
+        res.json({ message: "Deleted" });
+    } catch (err) {
+        res.status(500).json({ error: err.message });
+    }
 };
